move search icon || update imports || add to story
