--- ui/components/ui/search-icon/index.js
@@ -1,1 +1,0 @@
-export { default } from './search-icon.component';
