+++ ui/pages/settings/networks-tab/custom-content-search/custom-content-search.js
@@ -4,7 +4,7 @@ import Fuse from 'fuse.js';
 import InputAdornment from '@material-ui/core/InputAdornment';
 import TextField from '../../../../components/ui/text-field';
 import { I18nContext } from '../../../../contexts/i18n';
-import SearchIcon from '../../../../components/ui/search-icon';
+import SearchIcon from '../../../../components/ui/icon/search-icon';
 
 export default function CustomContentSearch({
   onSearch,
